01_ai_agents_first/03_litellm_openai_agent/01_litellm/myproject/pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "muhammadyasir678"}
]
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.73.0",
    "python-dotenv>=1.1.1",
]

[project.scripts]
gemini1 = "myproject.hello:gemini1"
gemini2 = "myproject.hello:gemini2"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
